Improvements to continent detection

+++ RandomVic2/utils/UtilLib.h
@@ -1,6 +1,7 @@
 #pragma once
 using namespace std;
 #include <algorithm>
+#include <iostream>
 
 static bool inRange(int offset, int high, int x)
 {
@@ -8,10 +9,10 @@ static bool inRange(int offset, int high, int x)
 }
 static double getDistance(int p1, int p2, int width, int height)
 {
-	const int x1 = p1 % width;
-	const int x2 = p2 % width;
-	const int y1 = p1 / height;
-	const int y2 = p2 / height;
+	const double x1 = p1 % width;
+	const double x2 = p2 % width;
+	const double y1 = (double)p1 / (double)height;
+	const double y2 = (double)p2 / (double)height;
 	return sqrt(((x1 - x2) *(x1 - x2)) + ((y1 - y2) *(y1 - y2)));
 }
